extract withdraw component and add storybook
