+++ app/components/layouts/LayoutAuthorized.tsx
@@ -4,7 +4,6 @@ import { Col, Row } from "reactstrap";
 import { NotificationWidget } from "../dashboard/notification-widget/NotificationWidget";
 import { DepositEthModal } from "../modals/DepositEthModal";
 import { TxSenderModal } from "../modals/txSender/TxSender";
-import { WithdrawModal } from "../modals/WithdrawModal";
 import { LayoutAuthorizedMenu } from "./LayoutAuthorizedMenu";
 
 import * as styles from "./LayoutAuthorized.module.scss";
@@ -21,7 +20,6 @@ export const LayoutAuthorized: React.SFC = ({ children }) => (
       </Row>
     </div>
     <DepositEthModal />
-    <WithdrawModal />
     <TxSenderModal />
   </>
 );
