+++ app/components/modals/txSender/Withdraw.stories.tsx
@@ -1,0 +1,6 @@
+import { storiesOf } from "@storybook/react";
+import * as React from "react";
+
+import { Withdraw } from "./Withdraw";
+
+storiesOf("Withdraw", module).add("default", () => <Withdraw />);
