+++ app/components/walletSelector/walletRecover/__tests__/WalletRecover.spec.e2e.ts
@@ -1,12 +1,12 @@
 import { tid } from "../../../../../test/testUtils";
 import {
   assertErrorModal,
+  assertLatestEmailSentWithSalt,
+  assertUserInDashboard,
   mockApiUrl,
   typeEmailPassword,
   typeLightwalletRecoveryPhrase,
 } from "../../../../e2e-test-utils";
-import { assertUserInDashboard } from "../../../../e2e-test-utils/index";
-import { assertLatestEmailSentWithSalt } from "./../../../../e2e-test-utils/index";
 
 describe("Wallet recover", () => {
   const words = [
